Extract browser features testing into own module
