+++ features/browser/presentation/src/androidTest/java/com/diskin/alon/appsbrowser/browser/di/TestInjector.java
@@ -1,4 +1,4 @@
-package com.diskin.alon.appsbrowser.browser.ui.di;
+package com.diskin.alon.appsbrowser.browser.di;
 
 import androidx.annotation.NonNull;
 import androidx.test.core.app.ApplicationProvider;
@@ -6,10 +6,10 @@ import androidx.test.core.app.ApplicationProvider;
 import com.diskin.alon.appsbrowser.browser.runner.TestApp;
 import com.diskin.alon.appsbrowser.browser.ui.BrowserFragmentTest;
 
-public class UiTestInjector {
+public class TestInjector {
 
     public static void inject(@NonNull BrowserFragmentTest test) {
-        UiTestAppComponent component = DaggerUiTestAppComponent.create();
+        TestAppComponent component = DaggerUiTestAppComponent.create();
         TestApp testApp = ApplicationProvider.getApplicationContext();
 
         component.inject(testApp);
