+++ settings.gradle
@@ -1,1 +1,5 @@
-include ':app', ':features:home', ':features:browser:presentation', ':features:common'
+include ':app',
+        ':features:home',
+        ':features:browser:presentation',
+        ':features:common',
+        ':features:browser:featureTest'
